% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simple_IIF_name_val.R
\name{create_simple_IIF_name_val}
\alias{create_simple_IIF_name_val}
\title{create a basic IIF calculation in either Tableau or PowerBi.}
\usage{
create_simple_IIF_name_val(calc_name, names_col, value_col, tool)
}
\arguments{
\item{calc_name}{name for the calculation.  This should be the same as}

\item{names_col}{column used as a name (e.g.  indicator or disaggregate)}

\item{value_col}{column used as a value}

\item{tool}{either Tableau or PowerBI}
}
\value{
calculations
}
\description{
create a basic IIF calculation in either Tableau or PowerBi.
}
\examples{
 \dontrun{
   create_simple_IIF_name_val(a_df)
}
}
