% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_iif_calc_tableau.R
\name{create_simple_indicator_val_twb}
\alias{create_simple_indicator_val_twb}
\title{create a basic IIF calculation in Tableau.}
\usage{
create_simple_indicator_val_twb(calc_name, names_col, value_col)
}
\arguments{
\item{calc_name}{name for the calculation.  This should be the same as}

\item{names_col}{column used as a name (e.g.  indicator or disaggregate)}

\item{value_col}{column used as a value}
}
\value{
calculations
}
\description{
create a basic IIF calculation in Tableau.
}
\examples{
 \dontrun{
   create_calc_template(csv_file)
}
}
