% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_simple_percentage_calc.R
\name{simple_percentage_calc}
\alias{simple_percentage_calc}
\title{Create percentage calculations with one numerator and one denominator}
\usage{
simple_percentage_calc(indicator, num, den, tool, table)
}
\arguments{
\item{indicator}{indicator name}

\item{num}{numerator}

\item{den}{denominator}

\item{tool}{either Tableau or PowerBI}

\item{table}{in PowerBI the table the num/dem belongs to}
}
\value{
a tibble of all calculations
}
\description{
Create percentage calculations with one numerator and one denominator
}
\examples{
 \dontrun{
   create_percentage_calc(a_df)
}
}
